requirements.txt
Django==2.0
django-cors-headers==2.1.0
django-extensions==1.9.9
django-rq==1.0.1
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
pymysql
